Fixed all import bugs (simplifying code).
