+++ decitala/database/corpora_models.py
@@ -7,39 +7,14 @@
 #
 # Location: NYC, 2021
 ####################################################################################################
-import json
-import os
-import natsort
-import unidecode
-
 from sqlalchemy import (
 	Column,
 	Integer,
 	String,
-	create_engine
 )
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker
-
-from music21 import converter
-from music21 import note
-
-here = os.path.abspath(os.path.dirname(__file__))
-decitala_path = os.path.dirname(os.path.dirname(here)) + "/corpora/Decitalas/"
-greek_path = os.path.dirname(os.path.dirname(here)) + "/corpora/Greek_Metrics/"
-prosody_path = os.path.dirname(os.path.dirname(here)) + "/corpora/Prosody/"
 
 Base = declarative_base()
-
-def get_engine(filepath, echo=False):
-	engine = create_engine(f"sqlite:////{filepath}", echo=echo)
-	Base.metadata.create_all(engine)
-	return engine
-
-def get_session(engine):
-	Session = sessionmaker(bind=engine)
-	session = Session()
-	return session
 
 class DecitalaData(Base):
 	"""
@@ -73,58 +48,3 @@ class ProsodicFragmentData(Base):
 	name = Column(String)
 	source = Column(String)
 	ql_array = Column(String)
-
-def _make_corpora_database(echo):
-	abspath_databases_directory = os.path.abspath("./databases/")
-	engine = get_engine(
-		filepath=os.path.join(abspath_databases_directory, "FRAGMENT_DATABASE.db"),
-		echo=echo
-	)
-	session = get_session(engine=engine)
-
-	for this_file in natsort.natsorted(os.listdir(decitala_path)):
-		# Will use the utils function eventually. Annoying bug.
-		split = this_file.split("_")
-		if len(split) == 2:
-			full_id = split[0]
-		elif len(split) >= 3:
-			if len(split[1]) == 1:  # e.g. ["80", "B", "..."]
-				full_id = "_".join([split[0], split[1]])
-			else:
-				full_id = split[0]
-
-		converted = converter.parse(os.path.join(decitala_path, this_file))
-		ql_array = json.dumps([x.quarterLength for x in converted.flat.getElementsByClass(note.Note)])
-		decitala = DecitalaData(
-			full_id=full_id,
-			name=this_file[:-4],
-			ql_array=ql_array
-		)
-		session.add(decitala)
-
-	for this_file in os.listdir(greek_path):
-		converted = converter.parse(os.path.join(greek_path, this_file))
-		ql_array = json.dumps([x.quarterLength for x in converted.flat.getElementsByClass(note.Note)])
-		greek_foot = GreekFootData(
-			name=this_file[:-4],
-			ql_array=ql_array
-		)
-		session.add(greek_foot)
-
-	for this_dir in os.listdir(prosody_path):
-		if this_dir == "README.md":
-			continue
-		subdir = os.path.join(prosody_path, this_dir)
-		for this_file in os.listdir(subdir):
-			converted = converter.parse(os.path.join(subdir, this_file))
-			ql_array = json.dumps([x.quarterLength for x in converted.flat.getElementsByClass(note.Note)])
-			prosodic_fragment = ProsodicFragmentData(
-				name=unidecode.unidecode(this_file[:-4]),
-				source=unidecode.unidecode(this_dir),
-				ql_array=ql_array
-			)
-			session.add(prosodic_fragment)
-
-	session.commit()
-
-# _make_corpora_database(echo=False)
